requirements.txt
fastapi
uvicorn
gunicorn
httpx
python-dotenv
python-multipart
